Log the store connection early in TPC. On PostgreSQL, this includes the backend PID.
Fixes #460
